# src/main/resources/swagger.yml
openapi: 3.0.3
info:
  title: VouckBack
  description: |-
    Microserviço para geração de voucher como descontos em valor ou percentual paras clientes em seu ponto de venda.

    _Se você que saber mais sobre a TIBBI, então  click [aqui](http://tibbi.com.br)._

  contact:
    email: mario7lagoas@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: http://localhost:8080/voucher-back/v1
...
paths:
  /voucher/filter:
    get:
      tags:
        - Voucher
      summary: Buscando Lista de Voucher Por Filtros
      description: Buscando Lista de Voucher Por Filtros na Plataforma
      operationId: buscandoListaFiltroVoucher
      parameters:
        - name: page
          description: "Numero da pagina atual"
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          description: "Quantidade de registro por pagina"
          in: query
          schema:
            type: integer
            default: 10
        - name: codigo
          description: "Código do Voucher disponibilizado"
          example: "240422172055905959002"
          in: query
          schema:
            type: string
            default: ""
        - name: descricao
          description: "Descrição da Promoção"
          example: "Promocao de Natal"
          in: query
          schema:
            type: string
            default: ""
        - name: clienteCpf
          description: "CPF do Cliente"
          example: "12345678900"
          in: query
          schema:
            type: string
            default: ""
        - name: pdv
          description: "Numero do PDV"
          example: "101"
          in: query
          schema:
            type: string
            default: ""
        - name: cupomResgate
          description: "Numero do cupom onde o voucher foi resgatado"
          example: "123456"
          in: query
          schema:
            type: string
            default: ""
        - name: voucherStatus
          description: "Status Atual do Voucher"
          example: "DISPONIBILIZADO"
          in: query
          schema:
            type: string
            default: ""
        - name: filialCnpj
          description: "Buscar pelo CNPJ da Filial"
          example: "11111111111111"
          in: query
          schema:
            type: string
            default: ""
        - name: tipoDesconto
          description: "Buscar pelo tipo de Desconto"
          example: "PERCENTUAL"
          in: query
          schema:
            type: string
            default: ""
        - name: inicio
          description: "Inicio da Promoção"
          example: "2024-05-12"
          in: query
          schema:
            type: string
            default: ""
        - name: fim
          description: "Fim da Promoção"
          example: "2029-05-31"
          in: query
          schema:
            type: string
            default: ""
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginacaoApiResponse'
                  - type: object
                    properties:
                      transacoes:
                        type: array
                        items:
                          $ref: '#/components/schemas/VoucherFiltroApiResponse'
  /voucher/consulta:
    post:
      tags:
        - Voucher
      summary: Consultando Promoções Disponiveis
      description: Consultando Promoções Disponiveis na Plataforma
      operationId: consultandoPromocoes
      requestBody:
        description: Consultando um Voucher
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsultaVoucherApiRequest'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultaVoucherApiResponse'
  /voucher/confirm:
    post:
      tags:
        - Voucher
      summary: Confirmando Voucher Disponível
      description: Confirmando Voucher Disponível na Plataforma
      operationId: confirmandoVoucher
      requestBody:
        description: Confirmando Voucher Disponível
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/VoucherApiRequest'
        required: true
      responses:
        '200':
          description: successful operation
  /voucher/cancel:
    post:
      tags:
        - Voucher
      summary: Cancelando Voucher Disponível
      description: Cancelando Voucher Disponível na Plataforma
      operationId: cancelandoVoucher
      requestBody:
        description: Cancelando Voucher Disponível
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/VoucherApiRequest'
        required: true
      responses:
        '200':
          description: successful operation
  /voucher/resgate:
    post:
      tags:
        - Voucher
      summary: Resgatando Um Voucher Disponivel
      description: Resgatando Um Voucher Disponivel na Plataforma
      operationId: resgatandoVoucher
      requestBody:
        description: Resgatando um Voucher
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoucherPromocaoApiRequest'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoucherPromocaoApiResponse'
        '404':
          description: not found operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: 404 NOT_FOUND
                  erros:
                    type: array
                    items:
                      properties:
                        codigo:
                          type: string
                          example: "Não encontrado."
                        mensagem:
                          type: string
                          example: "Voucher nao [240422172055905959002] encontrado"
        '422':
          description: unprocessable operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: 422 UNPROCESSABLE_ENTITY
                  erros:
                    type: array
                    items:
                      properties:
                        codigo:
                          type: string
                          example: "Não permitido."
                        mensagem:
                          type: string
                          example: "Desconto [8.99] maior que pagamento [5.99]"
  /voucher/consumer:
    post:
      tags:
        - Voucher
      summary: Consumindo Voucher Disponbilizado
      description: Consumindo Voucher Disponbilizado na Plataforma
      operationId: consumindoVoucher
      requestBody:
        description: Consumindo Voucher Disponbilizado
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoucherFinalizeApiRequest'
        required: true
      responses:
        '200':
          description: successful operation
  /voucher/rollback:
    post:
      tags:
        - Voucher
      summary: Estornando Voucher Disponbilizado
      description: Estornando Voucher Disponbilizado na Plataforma
      operationId: estornandoVoucher
      requestBody:
        description: Estornando Voucher Disponbilizado
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoucherFinalizeApiRequest'
        required: true
      responses:
        '200':
          description: successful operation
  /voucher/print:
    post:
      tags:
        - Voucher
      summary: Relatorio de Voucher
      description: Relatorio de Voucher na Plataforma
      operationId: relatorioVoucher
      requestBody:
        description: Imprimindo um relatorio de Voucher
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/VoucherFiltroApiResponse'
      responses:
        '200':
          description: successful operation
          content:
            text/plain:
              schema:
                type: string
                example: "data:application/pdf;base64,JVBERi0xLjUKJeLjz9MKMyAwIG9iago8PC9Db2xvclNwYWNlL0RldmljZVJHQi9TdWJ0eXBlL0ltYWdlL0hlaWdodCAzNTQvRmlsdGVyL0RDVERlY29kZS9UeXBlL1hPYmplY3QvV2lkdGggMjQ4MC9CaXRzUGVyQ29tcG9uZW50IDgvTGVuZ3RoIDM2ODg5Pj5zdHJlYW0K/9j/4AAQSkZJRgABAgAAZABkAAD/7AARRHVja3kAAQAEAAAACgAA/+4ADkFkb2JlAGTAAAAAAf/bAIQAFBAQGRIZJxcXJzImHyYyLiYmJiYuPjU1NTU1PkRBQUFBQUFERERERERERERERERERERERERERERERERERERERAEVGRkgHCAmGBgmNiYgJjZENisrNkREREI1QkRERERERERERERERERERERERERERERERERERERERERERERERERE/8AAEQgBYgmwAwEiAAIRAQMRAf/..."
...
components:
  schemas:
    PaginacaoApiResponse:
      properties:
        number:
          type: integer
          example: 0
          description: "Pagina atual"
        totalPages:
          type: integer
          example: 1
          description: "Total de paginas"
        totalElements:
          type: integer
          example: 1
          description: "Total de Items"
...
    VoucherFinalizeApiRequest:
      required:
        - transacao
      properties:
        transacao:
          type: string
          example: "aaaa-bbbb-cccc-dddd-eeeee"
          description: "GUID da Transação"
        valorPago:
          type: number
          example: 5.99
          description: "Valor Pago"
    VoucherApiRequest:
      required:
        - filialCnpj
        - clienteCpf
        - codigo
      properties:
        filialCnpj:
          type: string
          example: "11111111111111"
          description: "CNPJ da Loja"
        clienteCpf:
          type: string
          example: "12345678900"
          description: "CPF do Cliente"
        codigo:
          type: string
          example: "240422172055905959002"
          description: "Codigo do Voucher disponibilizado"
    ConsultaVoucherApiRequest:
      required:
        - filialCnpj
        - pdvFilial
        - clienteCpf
        - cupom
        - valorCompra
      properties:
        filialCnpj:
          type: string
          example: "11111111111111"
          description: "CNPJ da Loja"
        pdvFilial:
          type: string
          example: "101"
          description: "Numero do PDV"
        clienteCpf:
          type: string
          example: "12345678900"
          description: "CPF do Cliente"
        cupom:
          type: string
          example: "88888"
          description: "Numero do Cupom da Venda"
        valorCompra:
          type: number
          example: 10.99
          description: "Valor Total da Compra"
    ConsultaVoucherApiResponse:
      properties:
        status:
          type: string
          example: "OK"
          description: "Status da promoção"
        totalVoucher:
          type: integer
          example: 1
          description: "Total de Voucher disponibilizado"
        vouchers:
          type: array
          description: "Lojas associadas a promoção"
          items:
            $ref: '#/components/schemas/VoucherApiResponse'
    VoucherPromocaoApiRequest:
      required:
        - codigo
        - filialCnpj
        - pdv
        - cupom
        - valorPagamento
      properties:
        codigo:
          type: string
          example: "240422172055905959002"
          description: "Numero do Voucher"
        filialCnpj:
          type: string
          example: "11111111111111"
          description: "CNPJ da Loja"
        pdv:
          type: string
          example: "101"
          description: "Numero do PDV"
        cupom:
          type: string
          example: "88888"
          description: "Numero do Cupom da Venda"
        valorPagamento:
          type: number
          example: 5.99
          description: "Valor do Pagamento"
    VoucherPromocaoApiResponse:
      properties:
        status:
          type: string
          example: "OK"
          description: "Status do Resgate do Voucher"
        transacao:
          type: string
          example: "aaaa-bbbb-cccc-dddd-eeeee"
          description: "GUID de identificação da transação"
        descricao:
          type: string
          example: "Promocao de Natal"
          description: "Descrição da Promoção"
        valorDesconto:
          type: number
          example: 5.99
          description: "Valor do Desconto concedido"
    VoucherFiltroApiResponse:
      properties:
        guid:
          type: string
          example: "aaaa-bbbb-cccc-dddd-eeeee"
          description: "Identificador GUID"
        codigo:
          type: string
          example: "240422172055905959002"
          description: "Codigo do Voucher disponibilizado"
        clienteCpf:
          type: string
          example: "12345678900"
          description: "CPF do Cliente"
        filialCnpj:
          type: string
          example: "11111111111111"
          description: "CNPJ da Loja origem do voucher"
        promocaoGuid:
          type: string
          example: "aaaa-bbbb-cccc-dddd-eeeee"
          description: "Identificador GUID da promoção"
        descricao:
          type: string
          example: "Promocao de Natal"
          description: "Descrição da Promoção"
        pdv:
          type: string
          example: "102"
          description: "PDV origem do Voucher"
        cupom:
          type: string
          example: "123456"
          description: "Cupom origem do Voucher"
        pdvResgate:
          type: string
          example: "102"
          description: "PDV resgate do Voucher"
        filialCnpjResgate:
          type: string
          example: "11111111111111"
          description: "CNPJ da Loja resgate do voucher"
        cupomResgate:
          type: string
          example: "123456"
          description: "Cupom resgate do Voucher"
        valorPagamento:
          type: number
          example: 7.00
          description: "Valor pagamento"
        tipoDesconto:
          type: string
          example: "PERCENTUAL"
          description: "Tipo de desconto [PERCENTUAL/VALOR]"
        voucherStatus:
          type: string
          example: "DISPONIVEL"
          description: "Tipo de status do Voucher [DISPONIVEL - EM_USO - UTILIZADO - CANCELADO - EXPIRADO]"
        promocaoStatus:
          type: string
          example: "ATIVA"
          description: "Status Atual da promoção [ATIVA - PROGRESSO - BLOQUEADA - FINALIZADA] "
        valorDesconto:
          type: number
          example: 5.00
          description: "Valor do Desconto"
        valorMaximoDesconto:
          type: number
          example: 10.00
          description: "Valor maximo de desconto concedido"
        valorPago:
          type: number
          example: 7.00
          description: "Valor pago"
        fimResgate:
          type: string
          example: "2024-06-05 09:56:13"
          description: "Data e Hora fim do resgate"
        dataResgate:
          type: string
          example: "2024-05-14 09:56:13"
          description: "Data e Hora do resgate"
        inicio:
          type: string
          example: "2024-05-12 09:56:13"
          description: "Data e Hora inicio da Promoção"
        fim:
          type: string
          example: "2024-05-31 09:56:16"
          description: "Data e Hora fim da Promoção"
        dataAtualizacao:
          type: string
          example: "2024-05-14 09:56:13"
          description: "Data e Hora da ultima atualização"
        dataCadastro:
          type: string
          example: "2024-05-12 09:56:13"
          description: "Data e Hora do cadastrado"
    VoucherApiResponse:
      properties:
        descricao:
          type: string
          example: "Promocao de Natal"
          description: "Descrição da Promoção"
        codigo:
          type: string
          example: "240422172055905959002"
          description: "Codigo do Voucher disponibilizado"
        inicio:
          type: string
          example: "2024-05-12 09:56:13"
          description: "Data e Hora inicio da Promoção"
        fim:
          type: string
          example: "2024-05-31 09:56:16"
          description: "Data e Hora fim da Promoção"
        tipoDesconto:
          type: string
          example: "PERCENTUAL"
          description: "Tipo de desconto"
        promocaoStatus:
          type: string
          example: "ATIVA"
          description: "Status Atual da promoção da Promoção"
        valorDesconto:
          type: number
          example: 5.00
          description: "Valor do Desconto"
        valorMaximoDesconto:
          type: number
          example: 10.00
          description: "Valor maximo de desconto concedido"
...